fix: warnings and makefile

+++ src/parser/command_parser.c
@@ -1,4 +1,5 @@
 #include "command_parser.h"
+#include "parser.h"
 #include <stdbool.h>
 #include <stdio.h>
 #include <stdlib.h>
@@ -109,15 +110,15 @@ static struct parser_definition command_parser_def = {
     .start_state = COMMAND,
 };
 
-struct parser * command_parser_init() {
+struct parser *command_parser_init() {
     return parser_init(parser_no_classes(), &command_parser_def);
 }
 
-void command_parser_reset(struct parser * parser){
+void command_parser_reset(struct parser *parser) {
     parser_reset(parser);
-    struct parser_event * event = get_parser_event(parser);
-    event->type=MAY_VALID;
-    event->index=0;
+    struct parser_event *event = get_parser_event(parser);
+    event->type = MAY_VALID;
+    event->index = 0;
     event->arg1_len = 0;
     event->arg2_len = 0;
     event->cmd_len = 0;
@@ -127,7 +128,7 @@ struct parser_event *get_command(struct parser_event *event,
                                  struct parser *command_parser, char *buff,
                                  size_t count, size_t *nread) {
 
-    int i;
+    size_t i;
     for (i = 0; i < count && event->type == MAY_VALID; i++) {
         event = parser_feed(command_parser, buff[i]);
     }
@@ -135,10 +136,9 @@ struct parser_event *get_command(struct parser_event *event,
     return event;
 }
 
-struct parser_event * get_command_parser_event(struct parser * parser){
+struct parser_event *get_command_parser_event(struct parser *parser) {
     return get_parser_event(parser);
 }
-
 
 void command_parser_destroy(struct parser *parser) {
     if (parser != NULL)
